+++ controllers/sideloadController.js
@@ -1,8 +1,9 @@
-const DB = require('../lib/db/query');
+// const DB = require('../lib/db/query');
+const queries = require('../lib/db/queries');
 
 const getSideload = async (req, res) => {
   try {
-    const data = await DB.getSideload();
+    const data = await queries.getSideload();
     res.json(data);
   } catch (err) {
     console.log('Error: ', err);
