+++ controllers/testsController.js
@@ -7,7 +7,7 @@ const {
   removeRule,
   removeTarget,
 } = require('../lib/aws/eventBridgeActions');
-const DB = require('../lib/db/query');
+// const DB = require('../lib/db/query');
 const queries = require('../lib/db/queries');
 const { modelToEntityTest, entityToJsonTests, entityToJsonTest } = require('../mappers/test');
 const { modelToEntityTestRun } = require('../mappers/testRun');
@@ -118,19 +118,19 @@ const editTest = async (req, res) => {
   }
 };
 
-const getScheduledTests = async (req, res) => {
-  try {
-    const data = await DB.getTests();
-    res.json({ tests: data });
-  } catch (err) {
-    console.log('Error: ', err);
-  }
-};
+// const getScheduledTests = async (req, res) => {
+//   try {
+//     const data = await queries.getTests();
+//     res.json({ tests: data });
+//   } catch (err) {
+//     console.log('Error: ', err);
+//   }
+// };
 
 const getTest = async (req, res) => {
   try {
     const testId = req.params.id;
-    const data = await DB.getTest(testId);
+    const data = await queries.getTest(testId);
     res.json(data);
   } catch (err) {
     console.log('Error: ', err);
@@ -165,7 +165,7 @@ const runNow = async (req, res) => {
     const testId = req.params.id;
     // this url will need to come from a config file
     const lambdaURL = RULE_TARGET_INFO['test-route-packager'].url;
-    const data = await DB.getTestBody(testId);
+    const data = await queries.getTestBody(testId);
     axios.post(lambdaURL, data);
     res.status(200).send('OK');
   } catch (err) {
@@ -218,9 +218,9 @@ const getTestRun = async (req, res) => {
 const deleteTest = async (req, res) => {
   try {
     const testId = req.params.id;
-    const testName = await DB.getTestName(testId);
+    const testName = await queries.getTestName(testId);
     try {
-      await DB.deleteTest(testId);
+      await queries.deleteTest(testId);
     } catch (err) {
       console.log('Error: ', err);
       return err;
@@ -239,7 +239,6 @@ const deleteTest = async (req, res) => {
 
 exports.runNow = runNow;
 exports.createTest = createTest;
-exports.getScheduledTests = getScheduledTests;
 exports.getTest = getTest;
 exports.getTests = getTests;
 exports.getTestRuns = getTestRuns;
