+++ lib/db/queries/getTests.js
@@ -1,0 +1,10 @@
+const { dbQuery } = require('../conn');
+
+async function getTests() {
+  const query = 'SELECT * FROM tests';
+
+  const result = await dbQuery(query);
+  return result.rows;
+}
+
+module.exports = getTests;
