+++ lib/db/queries/index.js
@@ -6,6 +6,12 @@ const addTestAlerts = require('./addTestAlerts');
 const editTest = require('./editTest');
 const deactivateTestAssertions = require('./deactivateTestAssertions');
 const deleteTestAlerts = require('./deleteTestAlerts');
+const deleteTest = require('./deleteTest');
+const getSideload = require('./getSideload');
+const getTests = require('./getTests');
+const getTestName = require('./getTestName');
+const getTest = require('./getTest');
+const getTestBody = require('./getTestBody');
 
 // enables imports elsewhere under a common namespace
 //
@@ -22,4 +28,10 @@ module.exports = {
   editTest,
   deactivateTestAssertions,
   deleteTestAlerts,
+  deleteTest,
+  getSideload,
+  getTests,
+  getTestName,
+  getTest,
+  getTestBody,
 };
