+++ package.json
@@ -7,7 +7,8 @@
     "start": "node app.js",
     "dev": "nodemon app.js",
     "zip": "zip -r tests-crud.zip ./",
-    "lint": "./node_modules/.bin/eslint ."
+    "lint": "./node_modules/.bin/eslint .",
+		"test": "jest --verbose"
   },
   "engines": {
     "node": "^16.0.0"
@@ -27,6 +28,7 @@
   "devDependencies": {
     "eslint": "^8.19.0",
     "eslint-config-airbnb": "^19.0.4",
+    "jest": "^28.1.3",
     "nodemon": "^2.0.19"
   }
 }
