+++ routes/api.js
@@ -6,7 +6,6 @@ const { validateTest } = require('../validators/test');
 const router = express.Router();
 
 router.post('/tests', validateTest, testsController.createTest);
-// router.get('/tests', testsController.getScheduledTests);
 router.get('/tests', testsController.getTests);
 router.get('/tests/:id', testsController.getTest);
 router.get('/sideload', sideloadController.getSideload);
